requirements.txt
aiogram
aiogram-dialog
redis
python-dotenv
fluent.runtime
fluent.syntax
babel
openai
aiofiles
googletrans